+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 # 1.1.0
-- nfsw removed. Using vscode.FileSystemWatcher instead. With this Windows and Linux will be supported as well.
+- nfsw removed. Using vscode API instead. With this Windows and Linux will be supported as well.
+- 'Reload' button added.
 
 # 1.0.1
 - Readme: Comment added that the extension can only be used on macOS at the moment.